requirements.txt
yfinance
numpy
pandas
scikit-learn
tensorflow
matplotlib
streamlit
python-dotenv